proapi/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.17
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-xml==1.4.0
get==2018.11.19
httpie==1.0.2
idna==2.8
post==2018.11.20
public==2018.11.20
Pygments==2.3.1
pytz==2018.7
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
urllib3==1.24.1
